chore: upgrade to react-celo v4 (#68)
We'll need to upgrade to v5 of react-celo to keep revo working, as WC v1
will be sunset end of this month. I don't think v5 is ready yet, but
updating to use v4 hopefully makes it easier to do the upgrade when it's
available.

Key changes:
- I followed the steps in here
https://github.com/celo-org/react-celo/blob/master/guides/migrate-to-v4.md
- there was a lot of copy paste to rename things.
- the `@celo` packages needed to be updated to at least v2 to be
compatible with this change, but as far as i could see these upgrades
weren't breaking for this project
- `@celo/react-celo` no longer returns a full contract kit, i hacked at
this with a hook / function that returns a new contract kit to minimise
the changes.

I did a quick test and can still zap in and out / farm. the WC bottom
sheet in Valora doesn't display the correct dapp icon if you are testing
on localhost because the icon url contains `localhost:3000` but i'm
fairly sure this will work once deployed